Flytt fordeling av oppgaver ut av context slik at det kan gjenbrukes
